Update fish home module to user user config, and enable shell
The 'user' config was only used by this fish module, and became silly
as soon as I had to duplicate it to the git module. Instead I set it
once in the home config and reference it.

In addition this change includes enabling the fish shell. This makes it
possible to do things like set session variables, which I'll do in the
next commit.